backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-jose[cryptography]
pydantic
python-dotenv
datadog
psycopg2-binary
pytest
pytest-cov
httpx
pytest-dotenv
ddtrace
pydantic[email]
pydantic-settings
passlib[bcrypt]
pytest-asyncio 